[nvme] enforce 64bit offset for nvme_get_log_ext fn
Message-id: <1525888340-18642-68-git-send-email-dmilburn@redhat.com>
Patchwork-id: 214226
O-Subject: [RHEL7.6 BZ 1515584 67/75] nvme: enforce 64bit offset for nvme_get_log_ext fn
Bugzilla: 1515584
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Matias Bjrling <mb@lightnvm.io>

Compiling on 32 bits system produces a warning for the shift width
when shifting 32 bit integer with 64bit integer.

Make sure that offset always is 64bit, and use macros for retrieving
lower and upper bits of the offset.

(cherry picked from commit 7ec6074ff005e5f6cd2cf186a9ec7496c3db04f9)

Conflicts:
	drivers/nvme/host/core.c

RHEL differences:

Fixed up some white space.

+++ drivers/nvme/host/nvme.h
@@ -350,7 +350,7 @@ int nvme_delete_ctrl(struct nvme_ctrl *ctrl);
 int nvme_delete_ctrl_sync(struct nvme_ctrl *ctrl);
 
 int nvme_get_log_ext(struct nvme_ctrl *ctrl, struct nvme_ns *ns,
-		u8 log_page, void *log, size_t size, size_t offset);
+		u8 log_page, void *log, size_t size, u64 offset);
 
 struct sg_io_hdr;
